requirements.txt
Pillow
glob2